Refactoring: Rename Protocol classes with `_v4` postfix and move to protocol_v4 package
