+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/protocol_v4/SellerAsTakerProtocol_v4.java
@@ -15,30 +15,27 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.protocol.bisq_v1;
+package bisq.core.trade.protocol.bisq_v1.protocol_v4;
 
 
 import bisq.core.offer.Offer;
-import bisq.core.trade.model.bisq_v1.BuyerAsTakerTrade;
+import bisq.core.trade.model.bisq_v1.SellerAsTakerTrade;
 import bisq.core.trade.model.bisq_v1.Trade;
 import bisq.core.trade.protocol.TakerProtocol;
 import bisq.core.trade.protocol.TradeMessage;
-import bisq.core.trade.protocol.bisq_v1.messages.DelayedPayoutTxSignatureRequest;
-import bisq.core.trade.protocol.bisq_v1.messages.DepositTxAndDelayedPayoutTxMessage;
+import bisq.core.trade.protocol.bisq_v1.messages.CounterCurrencyTransferStartedMessage;
+import bisq.core.trade.protocol.bisq_v1.messages.DelayedPayoutTxSignatureResponse;
 import bisq.core.trade.protocol.bisq_v1.messages.InputsForDepositTxResponse;
-import bisq.core.trade.protocol.bisq_v1.messages.PayoutTxPublishedMessage;
+import bisq.core.trade.protocol.bisq_v1.messages.ShareBuyerPaymentAccountMessage;
 import bisq.core.trade.protocol.bisq_v1.tasks.ApplyFilter;
 import bisq.core.trade.protocol.bisq_v1.tasks.CheckIfDaoStateIsInSync;
 import bisq.core.trade.protocol.bisq_v1.tasks.TradeTask;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer.BuyerFinalizesDelayedPayoutTx;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer.BuyerProcessDelayedPayoutTxSignatureRequest;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer.BuyerSendsDelayedPayoutTxSignatureResponse;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer.BuyerSetupDepositTxListener;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer.BuyerSignsDelayedPayoutTx;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer.BuyerVerifiesPreparedDelayedPayoutTx;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer_as_taker.BuyerAsTakerCreatesDepositTxInputs;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer_as_taker.BuyerAsTakerSendsDepositTxMessage;
-import bisq.core.trade.protocol.bisq_v1.tasks.buyer_as_taker.BuyerAsTakerSignsDepositTx;
+import bisq.core.trade.protocol.bisq_v1.tasks.seller.MaybeCreateSubAccount;
+import bisq.core.trade.protocol.bisq_v1.tasks.seller.SellerCreatesDelayedPayoutTx;
+import bisq.core.trade.protocol.bisq_v1.tasks.seller.SellerSendDelayedPayoutTxSignatureRequest;
+import bisq.core.trade.protocol.bisq_v1.tasks.seller.SellerSignsDelayedPayoutTx;
+import bisq.core.trade.protocol.bisq_v1.tasks.seller_as_taker.SellerAsTakerCreatesDepositTxInputs;
+import bisq.core.trade.protocol.bisq_v1.tasks.seller_as_taker.SellerAsTakerSignsDepositTx;
 import bisq.core.trade.protocol.bisq_v1.tasks.taker.CreateTakerFeeTx;
 import bisq.core.trade.protocol.bisq_v1.tasks.taker.TakerProcessesInputsForDepositTxResponse;
 import bisq.core.trade.protocol.bisq_v1.tasks.taker.TakerPublishFeeTx;
@@ -56,22 +53,16 @@ import lombok.extern.slf4j.Slf4j;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
-public class BuyerAsTakerProtocol extends BaseBuyerProtocol implements TakerProtocol {
+public class SellerAsTakerProtocol_v4 extends BaseSellerProtocol_v4 implements TakerProtocol {
 
     ///////////////////////////////////////////////////////////////////////////////////////////
     // Constructor
     ///////////////////////////////////////////////////////////////////////////////////////////
 
-    public BuyerAsTakerProtocol(BuyerAsTakerTrade trade) {
+    public SellerAsTakerProtocol_v4(SellerAsTakerTrade trade) {
         super(trade);
-
         Offer offer = checkNotNull(trade.getOffer());
         processModel.getTradePeer().setPubKeyRing(offer.getPubKeyRing());
-    }
-
-    @Override
-    protected void onInitialized() {
-        super.onInitialized();
     }
 
 
@@ -80,31 +71,29 @@ public class BuyerAsTakerProtocol extends BaseBuyerProtocol implements TakerProt
     ///////////////////////////////////////////////////////////////////////////////////////////
 
     @Override
-    public void onMailboxMessage(TradeMessage message, NodeAddress peer) {
-        super.onMailboxMessage(message, peer);
+    public void onMailboxMessage(TradeMessage message, NodeAddress peerNodeAddress) {
+        super.onMailboxMessage(message, peerNodeAddress);
     }
 
 
     ///////////////////////////////////////////////////////////////////////////////////////////
-    // Take offer
+    // User interaction: Take offer
     ///////////////////////////////////////////////////////////////////////////////////////////
 
     @Override
     public void onTakeOffer() {
         expect(phase(Trade.Phase.INIT)
-                .with(TakerEvent.TAKE_OFFER))
+                .with(TakerEvent.TAKE_OFFER)
+                .from(trade.getTradingPeerNodeAddress()))
                 .setup(tasks(
                         CheckIfDaoStateIsInSync.class,
+                        MaybeCreateSubAccount.class,
                         ApplyFilter.class,
                         getVerifyPeersFeePaymentClass(),
                         CreateTakerFeeTx.class,
-                        BuyerAsTakerCreatesDepositTxInputs.class,
+                        SellerAsTakerCreatesDepositTxInputs.class,
                         TakerSendInputsForDepositTxRequest.class)
                         .withTimeout(120))
-                .run(() -> {
-                    processModel.setTempTradingPeerNodeAddress(trade.getTradingPeerNodeAddress());
-                    processModel.getTradeManager().requestPersistence();
-                })
                 .executeTasks();
     }
 
@@ -117,33 +106,36 @@ public class BuyerAsTakerProtocol extends BaseBuyerProtocol implements TakerProt
         expect(phase(Trade.Phase.INIT)
                 .with(message)
                 .from(peer))
-                .setup(tasks(TakerProcessesInputsForDepositTxResponse.class,
+                .setup(tasks(
+                        TakerProcessesInputsForDepositTxResponse.class,
                         ApplyFilter.class,
                         TakerVerifyAndSignContract.class,
                         TakerPublishFeeTx.class,
-                        BuyerAsTakerSignsDepositTx.class,
-                        BuyerSetupDepositTxListener.class,
-                        BuyerAsTakerSendsDepositTxMessage.class)
-                        .withTimeout(120))
-                .executeTasks();
-    }
-
-    protected void handle(DelayedPayoutTxSignatureRequest message, NodeAddress peer) {
-        expect(phase(Trade.Phase.TAKER_FEE_PUBLISHED)
-                .with(message)
-                .from(peer))
-                .setup(tasks(
-                        BuyerProcessDelayedPayoutTxSignatureRequest.class,
-                        BuyerVerifiesPreparedDelayedPayoutTx.class,
-                        BuyerSignsDelayedPayoutTx.class,
-                        BuyerFinalizesDelayedPayoutTx.class,
-                        BuyerSendsDelayedPayoutTxSignatureResponse.class)
+                        SellerAsTakerSignsDepositTx.class,
+                        SellerCreatesDelayedPayoutTx.class,
+                        SellerSignsDelayedPayoutTx.class,
+                        SellerSendDelayedPayoutTxSignatureRequest.class)
                         .withTimeout(120))
                 .executeTasks();
     }
 
     @Override
-    protected void handle(DepositTxAndDelayedPayoutTxMessage message, NodeAddress peer) {
+    protected void handle(DelayedPayoutTxSignatureResponse message, NodeAddress peer) {
+        super.handle(message, peer);
+    }
+
+    @Override
+    protected void handle(ShareBuyerPaymentAccountMessage message, NodeAddress peer) {
+        super.handle(message, peer);
+    }
+
+
+    ///////////////////////////////////////////////////////////////////////////////////////////
+    // Incoming message when buyer has clicked payment started button
+    ///////////////////////////////////////////////////////////////////////////////////////////
+
+    @Override
+    protected void handle(CounterCurrencyTransferStartedMessage message, NodeAddress peer) {
         super.handle(message, peer);
     }
 
@@ -153,28 +145,21 @@ public class BuyerAsTakerProtocol extends BaseBuyerProtocol implements TakerProt
     ///////////////////////////////////////////////////////////////////////////////////////////
 
     @Override
-    public void onPaymentStarted(ResultHandler resultHandler, ErrorMessageHandler errorMessageHandler) {
-        super.onPaymentStarted(resultHandler, errorMessageHandler);
+    public void onPaymentReceived(ResultHandler resultHandler, ErrorMessageHandler errorMessageHandler) {
+        super.onPaymentReceived(resultHandler, errorMessageHandler);
     }
 
 
     ///////////////////////////////////////////////////////////////////////////////////////////
-    // Incoming message Payout tx
-    ///////////////////////////////////////////////////////////////////////////////////////////
-
-    @Override
-    protected void handle(PayoutTxPublishedMessage message, NodeAddress peer) {
-        super.handle(message, peer);
-    }
-
-
-    ///////////////////////////////////////////////////////////////////////////////////////////
-    // Message dispatcher
+    // Massage dispatcher
     ///////////////////////////////////////////////////////////////////////////////////////////
 
     @Override
     protected void onTradeMessage(TradeMessage message, NodeAddress peer) {
         super.onTradeMessage(message, peer);
+
+        log.info("Received {} from {} with tradeId {} and uid {}",
+                message.getClass().getSimpleName(), peer, message.getTradeId(), message.getUid());
 
         if (message instanceof InputsForDepositTxResponse) {
             handle((InputsForDepositTxResponse) message, peer);
